Create a BSAExperiment object from a GATK varianttotable Table

This function reads a GATK varianttotable output table and a column data file to create a BSAExperiment object. The GATK VariantsToTable command should be used to generate the input table. See the GATK varianttotable documentation for details.

Usage

makeBSAExperimentFromGatkTable(
  gatk_table_path,
  col_data_path,
  drop_samples = c(),
  metadata = list()
)

Arguments

gatk_table_path

The path to the GATK table file.

col_data_path

Optional. If null, samplenames from the gatk table will be used to create a single column of metadata with the column name sample. However, additional sample metdata can be provided with this argument. If not NULL, this must be a character string specifying the path to the sample metadata. This must be a csv file with the extention .csv. The column sample must exist, and it must correspond to the sample names in the GATK varianttotable table.

drop_samples

A character vector of sample names to drop from the GATK table. Default is an empty character vector.

metadata

A list containing metadata for the BSAExperiment object. Recognized entries are population_structure, which may be 'F2' for an F2 population and 'RIL' for Recombinant Inbred Lines, population_1_n, which are the number of individuals in the 'low bulk' sample, and population_2_n, which are the number of individuals in the 'high bulk' sample. If population_2_n is not provided, it is equal to population_1_n by default.

Details

To create the input table for this function, use the following GATK VariantsToTable command:

gatk VariantsToTable \
   -V input.vcf \
   --split-multi-allelic \
   -F CHROM -F POS -F REF -F ALT -F MULTI-ALLELIC -F TYPE -F QUAL \
   -GF AD -GF DP -GF PL -GF GQ -GF GT \
   -O output.table

Examples

if (FALSE) {
gatk_table_path <- "path/to/gatk_table.table"
col_data_path <- "path/to/col_data.csv"
drop_samples <- c("sample1", "sample2")
metadata <- list(population_structure = 'RIL', population_1_n = 20)

bsa_experiment <- makeBSAExperimentFromGatkTable(gatk_table_path, col_data_path, drop_samples, metadata)
}