Process Chromosome Data

A helper function used by extract_alt_freq_dataframe to process data for' each chromosome separately. This function is designed to be used internally and is not exported.

Usage

process_chromosome(
  chr_sample_subset,
  depth_threshold,
  ref_freq_thres,
  hetero_freq_thres,
  alt_freq_thres,
  neighborhood_size,
  verbose
)

Arguments

depth_threshold

Integer specifying the minimum depth required for a variant to be considered. Default is 10.

ref_freq_thres

Numeric specifying the the maximum value for which a given variant will be 'rounded' to REF (0) for the smoothed allele frequency column

hetero_freq_thres

Numeric vector specifying the range of values for which a given variant will be 'rounded' to HET (0.5) for the smoothed allele frequency column

alt_freq_thres

Numeric specifying the minimum value for which a given variant will be 'rounded' to ALT (1) for the smoothed allele frequency column

neighborhood_size

Integer specifying the number of data points around a given variant that will be considered when determining if the variant is 'noisy', or not matching its immediate neighbors. Default is 20.

verbose

Logical indicating whether to print processing messages.

chr

Character string specifying the chromosome name.

bsae_sample_subset

A SummarizedExperiment object subset for a specific sample.

Value

A data frame with processed allele frequency data for the specified chromosome.